Included RAD on VF and MEasured comparison, excels of results for each setup
Now the journal reads also the results generated by the radiance file, merges the data so only matching times are used, and calculates the bifacial gain.
'
The excel files are the results for Measured (i.e. IMT1), Modeled with VF (mislabed as "IMT1_M") and Modeled with Radiance (labeled as "IMT1_Rad")